selfcross: regression tests, input validation, sites on the intersection
Review follow-up. The census kernel moves to a header shared by the tool
and a new core/test regression suite: clean/transverse/coplanar/grazing
fixtures, adjacency exclusion, canonical ordering, thread-count and
cell-size independence, maxedge gating, and input validation. A
nonpositive or non-finite --cell is refused at both the CLI and the
kernel, since a negative cell would empty every broad-phase bucket and
report a crossing surface clean. Crossing sites are now the midpoint of
the two triangles' shared intersection segment, a point both interiors
contain, rather than a centroid average that can sit off the crossing
for long triangles; verdicts, penetrations and angles are unchanged.
